[SCRUM-2] This is the base setup for the website we will make.
This uses node.js, express.js, react.js and a cloud hosted mongodb that has been made (so far it is empty).
You can connect via the CONNECTION_URL. This is not secure but for development purposes we will keep it like this. Let me know if it does not let you connect on the backend :).

Signed-off-by: MartiniMarkov <martinmarkovme@gmail.com>